+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+# 1.19
+* `correlationsum` now features the keyword `q` to calculate the q-order correlationsum.
+* Add fractal dimension estimation method by Molteno et al `molteno_dim`.
+
+# 1.18
+* `poincaresos` function now also works with input `Dataset` (and does linear interpolation between points sandwiching the dataset)
+
+# 1.17
+* `transit_time_statistics` deprecated in favor of `exit_entry_times`.
+* Added `mean_return_times` function for discrete systems.
+* Added `mean_return_times` function for continuous systems.
+
+# 1.16
+* `takens_best_estimate` now returns three arguments, the estimate plus the upper and lower 95%-confidence limits.
+
+# 1.15
+* New function `transit_time_statistics` that allows computing return times and transit times to subsets of the state space. (Currently for discrete systems only)
+* Moved support to Julia 1.5+.
+
 # 1.14
 * `orbitdiagram` now allows only collecting states within user-provided limits.
 # 1.13